Moved Control Fnc & Refactoring
